pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wrench-estimator"
version = "0.1.0"
description = "Mujoco-based end-effector wrench estimator for robotic arms"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = ["numpy", "mujoco>=2.3.7", "rerun-sdk>=0.30.0,<0.31"]

[tool.pixi.workspace]
channels = [
    "https://mirrors.tuna.tsinghua.edu.cn/anaconda/cloud/conda-forge/",
    "conda-forge",
]
platforms = ["linux-64"]

[tool.pixi.pypi-options]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
extra-index-urls = ["https://mirrors.aliyun.com/pypi/simple/"]

[tool.pixi.pypi-dependencies]
wrench-estimator = { path = ".", editable = true }
airbot-py = { path = "deps/airbot_py-5.1.6-py3-none-any.whl"}

[tool.pixi.feature.build.pypi-dependencies]
build = "*"

[tool.pixi.feature.publish.dependencies]
twine = "*"

[tool.pixi.environments]
build = ["build"]
publish = ["publish"]

[tool.pixi.tasks]
build = { cmd = "rm -rf dist && python -m build", default-environment = "build" }
publish-pypi = { cmd = "twine upload dist/*", depends-on = [
    "build",
], default-environment = "publish" }

[tool.setuptools]
packages = ["wrench_estimator"]
